refactor: vo 패키지 common으로 변경
