+++ src/main/java/kitchenpos/menu/domain/Product.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.vo.Name;
-import kitchenpos.menu.domain.vo.Price;
+import kitchenpos.common.vo.Name;
+import kitchenpos.common.vo.Price;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
