+++ src/test/java/kitchenpos/common/vo/QuantityTest.java
@@ -1,17 +1,17 @@
-package kitchenpos.menu.domain.vo;
+package kitchenpos.common.vo;
 
-import kitchenpos.menu.domain.vo.Name;
+import kitchenpos.common.vo.Quantity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-class NameTest {
+class QuantityTest {
 
     @DisplayName("[SUCCESS] 생성한다.")
     @Test
     void success_create() {
-        assertThatCode(() -> new Name("테스트용 이름"))
+        assertThatCode(() -> new Quantity(10))
                 .doesNotThrowAnyException();
     }
 }
