+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -4,8 +4,8 @@ import kitchenpos.config.ApplicationTestConfig;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.Product;
-import kitchenpos.menu.domain.vo.Name;
-import kitchenpos.menu.domain.vo.Price;
+import kitchenpos.common.vo.Name;
+import kitchenpos.common.vo.Price;
 import kitchenpos.dto.MenuCreateRequest;
 import kitchenpos.dto.MenuProductCreateRequest;
 import kitchenpos.dto.MenuResponse;
