+++ src/test/java/kitchenpos/menu/event/MenuHistoryEventListenerTest.java
@@ -7,9 +7,9 @@ import kitchenpos.menu.domain.MenuHistory;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProductHistories;
 import kitchenpos.menu.domain.Product;
-import kitchenpos.menu.domain.vo.Name;
-import kitchenpos.menu.domain.vo.Price;
-import kitchenpos.menu.domain.vo.Quantity;
+import kitchenpos.common.vo.Name;
+import kitchenpos.common.vo.Price;
+import kitchenpos.common.vo.Quantity;
 import kitchenpos.order.application.event.OrderPreparedEvent;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
